+++ app/src/main/java/com/example/jgjio_desktop/gostats/BarGraphFragment.java
@@ -17,15 +17,12 @@ import java.util.List;
 public class BarGraphFragment extends Fragment {
 
     public static final String EXTRA_LIST_ID = "com.example.jgjio_desktop.gostats.extra.LIST_ID";
-    private double mListID;
+    private int mListID;
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-
         View rootView = inflater.inflate(R.layout.graph_fragment, container, false);
-
-        mListID = getArguments().getDouble(EXTRA_LIST_ID);
-
+        mListID = getArguments().getInt(EXTRA_LIST_ID);
         populateGraph(rootView);
 
         return rootView;
@@ -33,9 +30,7 @@ public class BarGraphFragment extends Fragment {
 
     private void populateGraph(View view) {
         GraphView graphView = (GraphView) view.findViewById(R.id.graph);
-
         GraphViewModel vm = ViewModelProviders.of(this).get(GraphViewModel.class);
-
         BarGraphSeries<com.jjoe64.graphview.series.DataPoint> series = new BarGraphSeries<com.jjoe64.graphview.series.DataPoint>();
 
         vm.getList(mListID).observe(this, new Observer<List<DataPoint>>() {
@@ -56,16 +51,11 @@ public class BarGraphFragment extends Fragment {
         graphView.addSeries(series);
     }
 
-    public static BarGraphFragment newInstance(double listId) {
+    public static BarGraphFragment newInstance(int listId) {
         BarGraphFragment fragment = new BarGraphFragment();
         Bundle args = new Bundle();
-        args.putDouble(EXTRA_LIST_ID, listId);
+        args.putInt(EXTRA_LIST_ID, listId);
         fragment.setArguments(args);
         return fragment;
     }
-
-
-
-
-
 }
