+++ app/src/main/java/com/example/jgjio_desktop/gostats/HistogramGraphSettingsViewModel.java
@@ -2,10 +2,7 @@ package com.example.jgjio_desktop.gostats;
 
 import android.app.Application;
 import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.arch.paging.PagedList;
-
-import java.util.List;
+import android.util.Log;
 
 public class HistogramGraphSettingsViewModel extends AndroidViewModel {
     private AppRepository mRepository;
@@ -15,7 +12,7 @@ public class HistogramGraphSettingsViewModel extends AndroidViewModel {
         mRepository = new AppRepository(application);
     }
 
-    long getNumberOfDataPointsInList(double listID) {
+    long getNumberOfDataPointsInList(int listID) {
         return mRepository.getNumberOfDataPointsInList(listID);
     }
 }
